work Update - edit button now pulls in data from array list . Cant update or Delete Database yet

+++ app/src/main/java/com/example/pureflow_group_project/Reservoirs.java
@@ -4,9 +4,9 @@ public class Reservoirs {
     String name;
     double lat;
     double lon;
-    double lvl;
+    int lvl;
 
-    public Reservoirs(String name, double lat, double lon, double lvl) {
+    public Reservoirs(String name, double lat, double lon, int lvl) {
         this.name = name;
         this.lat = lat;
         this.lon = lon;
@@ -29,7 +29,7 @@ public class Reservoirs {
         return lon;
     }
 
-    public double getLvl() {
+    public int getLvl() {
         return lvl;
     }
